Rename timestampUtilities to index
